+++ website/src/stores/apps/tools.ts
@@ -1,0 +1,7 @@
+import {defineStore} from 'pinia';
+
+export const useToolsAppStore = defineStore('toolsApp', {
+    state: () => ({
+        search: '',
+    }),
+});